### 2025-06-30 #8513 @editorialbot
[PRE REVIEW]: visualkeras: A Python Package for Visualizing Keras and TensorFlow Models
**Submitting author:** @Soontosh (<a href="http://orcid.org/0009-0007-4780-2624">Santosh Varma Patapati Patapati</a>)
**Repository:** https://github.com/paulgavrikov/visualkeras/
**Branch with paper.md** (empty if default branch): JOSS-Submission
**Version:** v0.1.4
