### 2025-08-06 #8513 @abhishektiwari
Hi  @Soontosh I will be helping to edit this paper. Before we get started can you bring down the paper word count close to 1000 words. I see there are `Usage Examples` in the paper they are better served by package documentation/README (which is already the case).

Also please add contribution guideline and to mention/link to them in your README. You could for instance create a CONTRIBUTING.md file (see here for some examples: https://contributing.md/example/).
