### 2025-08-06 #8513 @abhishektiwari
@Soontosh code is contributing ~500 words.

Either keep one of the examples with smallest code and link code snippets used to generate others visualisation back to Github repository.

Or, you can link code snippets used to generate all visualisation back to Github repository.

Given APIs can evolve hence code in the paper can be outdated quickly.
