+++ paper.md
@@ -21,7 +21,7 @@ affiliations:
    index: 1
  - name: University of Mannheim
    index: 2
- - name: Cyrion
+ - name: Cyrion Labs
    index: 3
 date: 28 June 2025
 bibliography: paper.bib
@@ -62,195 +62,27 @@ Node coloring is fully customizable via the `color_map` parameter which maps lay
 Like the Layered View, the Graph View produces a Pillow `Image` object [@Clark2015] that can be displayed in Jupyter notebooks or saved to disk.
 
 # Usage Examples
-In this section, we provide examples of how to use visualkeras to visualize Keras models in different ways. The examples shown in Figure 1, Figure 2, Figure 3, and Figure 4 demonstrate the flexibility and customization options available in the package. The generated graphical visualizations can be displayed in Jupyter notebooks or saved as image files for use in publications or presentations.
+/*In this section, we provide examples of how to use visualkeras to visualize Keras models in different ways. The examples shown in Figure 1, Figure 2, Figure 3, and Figure 4 demonstrate the flexibility and customization options available in the package. The generated graphical visualizations can be displayed in Jupyter notebooks or saved as image files for use in publications or presentations.*/
+
+In this section, we provide examples of Keras model visualizations that were generated using visualkeras. The examples shown in Figure 1, Figure 2, Figure 3, and Figure 4 demonstrate the flexibility and customization options available in the package. Code snippets used to generate these examples are written in the visualkeras GitHub repository's [examples.md file](LINK). Generated graphical visualizations can be displayed inline in Jupyter notebooks or saved as image files for use in publications or presentations.
 
 ## Layered View
 
 ### Basic Usage
-```python
-import visualkeras
-from tensorflow.keras.models import Sequential
-from tensorflow.keras.layers import Dense, Conv2D, MaxPooling2D, Flatten, Dropout
-
-# Define a simple sequential model
-simple_sequential_model = Sequential([
-    Conv2D(32, (3, 3), activation='relu', input_shape=(28, 28, 1)),
-    MaxPooling2D((2, 2)),
-    Conv2D(64, (3, 3), activation='relu'),
-    MaxPooling2D((2, 2)),
-    Conv2D(64, (3, 3), activation='relu'),
-    Flatten(),
-    Dense(64, activation='relu'),
-    Dropout(0.5),
-    Dense(10, activation='softmax')
-])
-
-# Basic usage of visualkeras
-basic_layered_img = visualkeras.layered_view(simple_sequential_model, draw_funnel=False)
-
-# Display the image
-basic_layered_img.show()
-```
 
 ![An example of layered style visualization on a simple sequential model with little styling](joss_figures/figure1_layered_basic.png)
 
 ### Advanced Usage
-```python
-import visualkeras
-from tensorflow.keras.models import Model
-from tensorflow.keras.layers import Dense, Conv2D, MaxPooling2D, Flatten, Dropout
-from PIL import ImageFont
-from collections import defaultdict
-
-# Define custom font for the model visualization
-custom_font = ImageFont.truetype("arial.ttf", 24)
-
-# Define custom color map
-color_map = defaultdict(dict)
-color_map[Conv2D]['fill'] = 'orange'
-color_map[MaxPooling2D]['fill'] = 'red'
-color_map[Dense]['fill'] = 'teal'
-
-# Define a larger sequential model with more complexity
-complex_sequential_model = Sequential([
[…]
