### 2025-08-07 2d276f9
Added usage examples as markdown file
