### 2025-08-07 a6cc0da
Added figures to useage_examples.md and added associated figures to the figures folder
