### 2025-08-07 e9f677b
Made contribution guidelines in CONTRIBUTING.md
