+++ CONTRIBUTING.MD
@@ -1,0 +1,94 @@
+# Contributing to visualkeras
+First off, thank you for your interest in contributing to visualkeras! ❤️
+
+All types of contributions are encouraged and valued. See the [Table of Contents](#table-of-contents) for different ways to help and details about how this project handles them. Please make sure to read the relevant section before making your contribution. It will make it a lot easier for the maintainer and smooth out the experience for all involved. The community looks forward to your contributions!
+
+> And if you like the project, but just don't have time to contribute, that's fine. There are other easy ways to support the project and show your appreciation, which we would greatly appreciate:
+> - Star the project on GitHub
+> - [Cite the repository in your work](https://github.com/paulgavrikov/visualkeras/blob/master/CITATION.cff)
+> - Tweet about it
+> - Refer this project in your project's readme
+> - Share it with your friends
+
+## Table of Contents
+
+- [I Have a Question](#i-have-a-question)
+- [Reporting Bugs](#reporting-bugs)
+- [Suggesting Features and Improvements](#suggesting-features-and-improvements)
+- [I Want To Contribute](#i-want-to-contribute)
+- [Your First Code Contribution](#your-first-code-contribution)
+- [Improving The Documentation](#improving-the-documentation)
+
+## I Have a Question
+
+> If you want to ask a question, we assume that you have already read the [README](https://github.com/paulgavrikov/visualkeras/blob/master/README.md) and its FAQ section.
+
+Before you ask a question, it is best to search for [existing issues](https://github.com/paulgavrikov/visualkeras/issues?q=is%3Aissue) that might help you. In case you have found a suitable issue and still need clarification, you can write your question and @Soontosh in the comments of that issue.
+
+If you then still feel the need to ask a question and need clarification, we recommend the following:
+- Open an [Issue](https://github.com/paulgavrikov/visualkeras/issues).
+- Provide as much context as you can about what you're running into.
+- Provide project and platform versions (Python, Keras, visualkeras, etc.), depending on what seems relevant.
+
+## Reporting Bugs
+
+If you come across a bug while using visualkeras, please open a new issue. You can use our *bug report template* to help structure your issue. In your issue, include the following details:
+- A clear description of the bug.
+- Minimal code and steps to reproduce the bug
+- Expected behavior
+- If applicable, a screenshot or example of the bug.
+- Environment (OS, Python version, visualkeras version or commit hash, Keras version, etc.)
+- Any additional context that might help us understand the issue.
+
+
+## Suggesting Features and Improvements
+
+To suggest new features or suggest improvements to existing features, please open a new issue. You can use our *feature request template* to help structure your suggestion. In your issue, include the following details:
+- Whether or not your feature reqeust is related to a problem
+- A description of the solution you'd like to see and what you want to happen
+- A description of alternatives you've considered
+- Any additional context that might help us understand the feature or improvement
+
+## I Want To Contribute
+
+If you want to contribute, please take a look at [existing issues](https://github.com/paulgavrikov/visualkeras/issues) and see if you can help with any of them, in which case you can comment on the issue to express your interest. If you have a new idea, please open an issue to discuss it first. This helps us avoid duplicate work and ensures that your contribution aligns with the project's goals.
+
+We also have a list of improvements that we would like to see in the project, which you can find in the project [README](https://github.com/paulgavrikov/visualkeras/blob/master/README.md).
+
+
+## Your First Code Contribution
+
+To get started with development, clone the repository and install the required dependencies:
+
+```bash
+git clone https://github.com/yourusername/visualkeras.git
+cd visualkeras
+pip install -r requirements.txt
+```
+
+Make sure to run the tests under the `tests/` folder to ensure everything is working correctly.
+
[…]
