### 2025-08-07 #8513 @Soontosh
Hello @abhishektiwari .

I've updated the JOSS-Submission branch (https://github.com/paulgavrikov/visualkeras/tree/JOSS-Submission) as you suggested. I've removed the code snippets and instead just included sample figures, linking to the code examples in the main branch. I've also creating a CONTRIBUTING.md in the main branch.
