### 2025-08-07 9a15272
Adjust PIL BibTeX reference to be more professional, remove duplicate Usage Examples section, fix typo, attempt to addrress L49 overflow
