### 2025-08-24 #8774 @abhishektiwari
@preetishkakkar, @sampottinger Thanks for agreeing to review this JOSS paper. I see you have not created your reviewer checklist yet. Please run this command in a separate comment to create the checklist:

To provide feedback to author please create a new issue in the [target repository](https://github.com/vinayakdasgupta/anvay) and link to those issues (especially acceptance-blockers) by leaving comments in the review thread below.
