### 2025-08-30 paulgavrikov/visualkeras#90 @sampottinger
[REQUEST] README issue links / graph view example
This is not related to an issue in production code. This highlights opportunities for expansion of the README.

**Describe the solution you'd like**

I have two suggestions about further improving the README described below.

> Usage

Please strongly consider adding a graph view example. It may be prudent to expand your example to include the model being visualized as well.

> We plan to add support for Keras version 1

You may want to link to issues if discussing planned work in the README.

You could also move this to a documentation website though the README seems appropriate.

**Additional context**

This is part of https://github.com/openjournals/joss-reviews/issues/8774.
