### 2025-08-30 paulgavrikov/visualkeras#91 @sampottinger
[REQUEST] Developer guidelines
Not related to an issue in production code. Simply a request that may be able to help further improve the new developer experience.

**Describe the solution you'd like**

`CONTRIBUTING.md` looks great! However, there may be opportunities to offer additional documentation about how to properly get started as a new developer.

First, it may be best to indicate how to perform development, either here or in README.md. This may include indication if any additional libraries are recommended for development since you only have install minimums in setup.py and what commands you expect folks to check (and how to run them) before submitting a PR. For example, for linting and unit tests before they submit a PR.

Second, I don't see anything about code style guide or lint guidelines. At minimum, please consider referencing and ideally checking the [PEP 8](https://peps.python.org/pep-0008/) rules. Furthermore, you are also using docstrings in some places that seem to follow a specific format (sphinx?). Which format is that and where can folks learn more about it to ensure their contributions are well aligned with your vision for the software? Though it may feel obvious, there are a number of competing standards and providing this information can help ensure a positive new contributor experience, especially for those earlier in their Python journey.

You may create a separate `DEVELOPING.md` but `CONTRIBUTING.md` is short enough it may be added directly.

**Additional context**

Part of https://github.com/openjournals/joss-reviews/issues/8774
