### 2025-08-30 paulgavrikov/visualkeras#92 @sampottinger
[REQUEST] Further document tests
This is not an issue in production code but further documentation of your testing approach may help new developers work within your project and expand test coverage when adding or updating functionality.

**Describe the solution you'd like**

I think there is some growth potential to further document how the tests work. I think you are doing a great job testing a diversity of architectures but it is a little unclear to me what is being asserted and where.

I think there is a great opportunity to add a `README.md` to `visualkeras/tests` that describes to other developers / contributors how the tests work, what assertions you expect, and what your project standards are for new code to be considered tested!

This could be added to `CONTRIBUTING.md` or `DEVELOPING.md`.

**Additional context**

This is part of https://github.com/openjournals/joss-reviews/issues/8774
