### 2025-08-30 paulgavrikov/visualkeras#93 @sampottinger
[REQUEST] Further document layer_untils
Though production code, this is a request for further documentation to help future contributors and not a known functionality issue. That said, there may be opportunities to expand throwing exceptions as it is unclear if some conditional branches may cause the library to silently enter into an undefined state.

**Describe the solution you'd like**

Looking at `layer_utils.py`, I like the functional decomposition in here but some of this logic is doing some advanced things like inspection. It might be ideal to add docstrings to these methods as well to help folks read your code. For example, I think I understand what the purpose of `find_layer_by_id` is but it might be helpful to say more about when you expect each branch to trigger to help those who are reviewing your code but are less familiar with the the internals of some of these other libraries. Finally, there is some branching where I'm unclear if your library is entering into an invalid state. Taking all of these thoughts together, consider something like:

```
def get_layers(model):
	"""Get the layers from a model depending on its construction.

	Retrieve the layers from a model depending on its construction. For example,
	models constructed through sequential or functional APIs would have
	`_layers` whereas subclasses may put them on attributes.

	:param model: The keras or tensorflow model from which layers should be
		returned.
	:return: The layers if found or rasies exception otherwise.
	"""
    is_sequential = hasattr(model, '_layers')
    if is_sequential:
	    return model._layers

    is_attribute_based = hasattr(model, '_self_tracked_trackables')
    if is_attribute_based:
        return model._self_tracked_trackables

    raise RuntimeError('Model given without _layers or _self_tracked_trackables.')
```

You could provide additional documentation through supporting module-level documentation but I think inline docs and potentially a little more decomposition is appropriate.

**Additional context**

This is part of https://github.com/openjournals/joss-reviews/issues/8774.
