### 2025-08-30 paulgavrikov/visualkeras#94 @sampottinger
[REQUEST] Package-level documentation available via help
I love the docstrings you have provided. However, there is very limited help available to users doing `help` on your package itself. This may be an opportunity to further expand documentation to helpful folks get started.

**Describe the solution you'd like**

I think it is great that you are putting these elements at the package level. However, you should also consider supporting developers who are trying to `help()` your package by adding a package-level docstring. Note the guidance from [PEP 0257](https://peps.python.org/pep-0257/) that "all modules should normally have docstrings ...." At minimum though, doing this in your `__init__.py` would be helpful.

A package website could be helpful instead! However, if sticking to just keeping your docstrings in source, supporting `help` is a good idea which I recommend to include package-level information.

**Additional context**

Part of https://github.com/openjournals/joss-reviews/issues/8774.
