### 2025-08-30 paulgavrikov/visualkeras#95 @sampottinger
[REQUEST] Make Travis available for contributors
Though I am thankful to `CONTRIBUTORS.md`, it is a little hard to tell if I am passing tests in a clean environment that would be acceptable ahead of a pull request. How might contributors interact with your CI / CD system to check their work? What if tests pass on their machine but not in travis, how would they debug the problem?

**Describe the solution you'd like**

I don't see build status reported on PRs. Do you have a link to public builds? It would be great if those could be accessed by contributors. Also, consider using a status badge.

You could consider GitHub actions or others but public build status should be strongly considered.

**Additional context**

Part of https://github.com/openjournals/joss-reviews/issues/8774
