### 2025-08-30 #8774 @sampottinger
Hello again @Soontosh!

Thank you for your patience and a great package. I have left my first review above with links to issues opened in your repo per JOSS guidelines. Please review the notes section above and please be aware that I did not open issues for the paper as I could not find the paper file in the repository.

Please let me know what you think of this feedback and if I can be of any further assistance. I look forward to continuing to work together on this very useful addition to the community!

With gratitude,
Sam

CC @abhishektiwari
