### 2025-09-02 34c20c7
Added usage example for graph-based view to README
