### 2025-09-02 9051226
Set height limit for larger figures in README, also reorder Graph_View usage examples
