### 2025-09-02 paulgavrikov/visualkeras#94 @Soontosh
Will add `help()`!
