### 2025-09-02 63ceb0a
Added clarifications around formatting for CONTRIBUTING.md
