### 2025-09-02 paulgavrikov/visualkeras#92 @Soontosh
Addressed in PR #98
