### 2025-09-02 78fda1c
Added a README for tests and referenced in CONTRIBUTING
