### 2025-09-02 paulgavrikov/visualkeras#93 @Soontosh
Fixed in #99
