### 2025-09-13 7e1595c
Refactor layer utilities for Keras compatibility
- Introduced .gitignore to exclude common Python artifacts and build files.
- Updated `graph_view` and `layered_view` functions to improve compatibility with Keras 2 and 3.
- Added helper functions to retrieve layer input and output shapes robustly.
- Applied a patch to ensure single-item outputs are unwrapped correctly in Keras models.
