+++ visualkeras/layer_utils.py
@@ -2,6 +2,7 @@ import numpy as np
 from .utils import get_keys_by_value
 from collections.abc import Iterable
 import warnings
+from typing import List, Tuple, Union, Optional
 
 try:
     from tensorflow.keras.layers import Layer
@@ -74,9 +75,18 @@ def get_incoming_layers(layer):
 
 
 def get_outgoing_layers(layer):
-    """Yield outgoing (child) layers for a given layer."""
+    """Yield outgoing (child) layers for a given layer.
+
+    Supports both legacy Node API (TF/Keras <= 2.15) and the new Node API
+    (TF >= 2.16 / Keras >= 3).
+    """
     for i, node in enumerate(layer._outbound_nodes):
-        yield node.outbound_layer
+        if hasattr(node, 'outbound_layer'):
+            # Old Node API
+            yield node.outbound_layer
+        else:
+            # New Node API (Keras 3): node.operation is the target layer
+            yield node.operation
 
 
 def model_to_adj_matrix(model):
@@ -304,6 +314,176 @@ def is_internal_input(layer):
         pass
 
     return False
+
+
+# ----------------------------
+# Shape utilities (Keras 2/3)
+# ----------------------------
+
+def _tensor_shape_to_tuple(shape_obj) -> Optional[Tuple]:
+    """Convert TensorShape/KerasTensor.shape to a Python tuple of ints/None.
+
+    Returns None if conversion is not possible.
+    """
+    if shape_obj is None:
+        return None
+    # TensorFlow TensorShape has as_list
+    if hasattr(shape_obj, 'as_list'):
+        try:
+            return tuple(shape_obj.as_list())
+        except Exception:
+            pass
+    # Otherwise assume iterable of dims
+    try:
+        dims = []
+        for d in shape_obj:
+            # Some dims are Dimension-like; try int() with fallback to None
+            if d is None:
+                dims.append(None)
+            else:
+                try:
+                    dims.append(int(d))
+                except Exception:
+                    dims.append(None)
+        return tuple(dims)
+    except TypeError:
+        return None
+
+
+def get_layer_output_shape(layer) -> Union[Tuple, List[Tuple], None]:
+    """Best-effort retrieval of a layer's output shape as tuple(s).
+
+    Works across Keras/TF versions where `.output_shape` might not be present
+    (e.g., Keras 3 InputLayer).
+    """
+    # 1) Direct attribute (older versions)
+    s = getattr(layer, 'output_shape', None)
+    if s is not None:
+        return s
+
+    # 2) From `output` tensor(s)
+    out = getattr(layer, 'output', None)
+    if out is not None:
+        if isinstance(out, (list, tuple)):
+            shapes = [_tensor_shape_to_tuple(t.shape) for t in out]
+            return shapes
+        else:
+            return _tensor_shape_to_tuple(out.shape)
+
+    # 3) Fallbacks for Input-like layers
+    for attr in ('batch_shape', 'batch_input_shape', 'input_shape', 'shape'):
+        s = getattr(layer, attr, None)
+        if s is not None:
+            # Ensure tuple(s)
+            if isinstance(s, (list, tuple)) and len(s) > 0 and isinstance(s[0], (list, tuple)):
+                return [tuple(x) for x in s]
+            if hasattr(s, 'as_list'):
+                try:
+                    return tuple(s.as_list())
+                except Exception:
+                    pass
+            # Single tuple
+            if isinstance(s, (list, tuple)):
+                return tuple(s)
+            # Unknown format
+            break
+    return None
+
+
+def get_layer_input_shape(layer) -> Union[Tuple, List[Tuple], None]:
+    """Best-effort retrieval of a layer's input shape as tuple(s)."""
+    # 1) Direct attribute
[…]
